+++ test/ssl/TestSSL.java
@@ -9,7 +9,8 @@ package ssl;
  *
  * @author rasto
  */
-import color.SPD1;
+import light.implementations.SimpleSpotLight;
+import color.implementations.SPD1;
 import color.SpectralPowerDistribution;
 import light.*;
 import java.io.*;
@@ -24,15 +25,15 @@ public class TestSSL {
         long occurencesY[] = new long[360 / step];
         long occurencesL[] = new long[800 / step];
         
-        float[] v3 = {0,0,0};  
+        double[] v3 = {0,0,0};  
         SpectralPowerDistribution spd = new SPD1();
-        SimpleSpotLight ssl = new SimpleSpotLight(spd, new float[]{0,0,0}, new float[]{1,1,1},10);
+        SimpleSpotLight ssl = new SimpleSpotLight(spd, new double[]{0,0,0}, new double[]{1,1,1},10);
         
         try{
         FileWriter fw = new FileWriter("test/ssl/sslAnglesXY.txt");
         
         for(long a = 0;a < numberofbeams;++a){
-                float[] b = ssl.getNextBeam();
+                double[] b = ssl.getNextBeam();
                 //this part is to create histogram
                 occurencesX[(int)floor( b[3]/ step)]++; 
                 occurencesY[(int)floor( b[4]/ step)]++;
